sw/librespresso/requirements.txt
pydantic
pyserial
pytest